We changed the telemetry and removed the pinpoint drive and uncommented the limelight file
